M7-T2 (run 20): SPHOME implemented + hw_verified — verbatim float home (negative/fractional, no clamp); void coord->errnum 8; home_x/y widened i32->f64

+++ crates/sb-render/src/compositor.rs
@@ -254,7 +254,7 @@ fn render_sprite(sp: &Sprite, sheet: &GrpPage, fb: &mut Framebuffer) {
     if sp.scale_x == 0.0 || sp.scale_y == 0.0 {
         return;
     }
-    let (hx, hy) = (sp.home_x as f64, sp.home_y as f64);
+    let (hx, hy) = (sp.home_x, sp.home_y);
     let fwd = |lx: f64, ly: f64| {
         let ax = (lx - hx) * sp.scale_x;
         let ay = (ly - hy) * sp.scale_y;
